+++ NEWS.md
@@ -22,6 +22,7 @@ BUGFIXES :
 * Fix `createDSR()` in API mode : daily binding constraint takes 366 rows.
 * Fix `createCluster()` and `editCluster()` parameter `list_pollutants` stop if Antares Version < 8.6.0
 * `getJobs()` no longer returns duplicates and displays the two new columns `owner_id` and `owner_name`.
+* `updateScenarioBuilder()` works for NTC part : allow cartesian in the merge.
 
 # antaresEditObject 0.6.1